Big cleanup + remove imagej/imglib2 dependencies. NotebookService is working with two generic methods.
