--- src/main/java/net/imagej/notebook/displayer/StringDisplayer.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2017 SciJava.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package net.imagej.notebook.displayer;
-
-import java.util.HashMap;
-import java.util.Map;
-import jupyter.Displayer;
-
-public class StringDisplayer extends Displayer {
-
-    private static final StringDisplayer INSTANCE = new StringDisplayer();
-
-    public static Displayer<Object> get() {
-        return INSTANCE;
-    }
-
-    @Override
-    public Map<String, String> display(Object obj) {
-        Map<String, String> result = new HashMap<>();
-
-        result.put("text/plain", obj.toString());
-
-        return result;
-    }
-
-}
